+++ src/components/db/richtext.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import type {
   RichTextPropertyItemObjectResponse,
-} from '../../types'
+} from '../../server/types'
 import TextBlock from '../page/text'
 
 export type RichTextProps = {
