+++ src/components/page/callout.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import TextBlock from './text'
 import type {
   CalloutBlockObjectResponse,
-} from '../../types'
+} from '../../server/types'
 
 export type CalloutBlockProps = {
   block: CalloutBlockObjectResponse
